requirements.txt
gradio==5.35.0
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.3.0
ultralytics==8.3.162